refactor: move out the loadTemplateConfig from Gigantic generator.js
