+++ apps/generator/test/generator.test.js
@@ -131,7 +131,7 @@ describe('Generator', () => {
       util = require('../lib/utils');
       filtersRegistry = require('../lib/filtersRegistry');
       hooksRegistry = require('../lib/hooksRegistry');
-      templateConfigValidator = require('../lib/templateConfigValidator');
+      templateConfigValidator = require('../lib/templateConfig/validator.js');
       xfsMock = require('fs.extra');
       const { AsyncAPIDocument } = require('@asyncapi/parser/cjs/models/v2/asyncapi');
       asyncApiDocumentMock = new AsyncAPIDocument({ 'x-parser-api-version': 0 });
